Analytics: promisify the refreshCountryCodeCookieGdpr function and move it to utils
Make it return a promise, replacing the continuation-callback-passing convention. The
new usage is arguably simpler.
