+++ client/lib/analytics/utils/index.js
@@ -14,3 +14,4 @@ export { default as urlParseAmpCompatible } from './url-parse-amp-compatible';
 export { default as shouldReportOmitBlogId } from './should-report-omit-blog-id';
 export { default as saveCouponQueryArgument } from './save-coupon-query-argument';
 export { default as getNormalizedHashedUserEmail } from './get-normalized-hashed-user-email';
+export { default as refreshCountryCodeCookieGdpr } from './refresh-country-code-cookie-gdpr';
